+++ tests/testthat/test-liana_plot.R
@@ -7,11 +7,13 @@ liana_cpdb <- readRDS(file.path(liana_path, "testdata",
 test_that("Test liana dotplot", {
     exp1 <- readRDS(file.path(liana_path, "testdata",
                               "output", "liana_dotplot_out.RDS"))
-    res1 <- liana_dotplot(liana_cpdb,
+    res1 <- liana_dotplot(liana_cpdb %>% # invert
+                              mutate(pvalue = -log10(pvalue+0.00000001)),
                           source_groups = "B",
                           target_groups = c("NK", "CD8 T"),
                           magnitude = "lr.mean",
-                          specificity = "pvalue")
+                          specificity = "pvalue",
+                          show_complex = TRUE)
 
     expect_identical(res1$data, exp1$data)
 })